Fix execution of ws in tests

+++ defis/1/la_prosperite_des_charlatans/tests/test1/test.py
@@ -34,15 +34,17 @@ if __name__ == "__main__":
             changes_output = test_path + "output/changes.txt"
 
             # CLEAN OUTPUT
-            os.remove(view0_output)
-            os.remove(changes_output)
+            if os.path.isfile(view0_output):
+                os.remove(view0_output)
+
+            if os.path.isfile(changes_output):
+                os.remove(changes_output)
 
             cmd = [group_path + EXEC_PATH, test_path + "input/data.csv", test_path + "input/user.txt", view0_output, changes_output]
             out = check_output(cmd, stderr=STDOUT, timeout=30).decode("utf-8")
 
             changes_expected = test_path + EXPECTED_PATH + "changes_expected.txt"
             view0_expected = test_path + EXPECTED_PATH + "view0_expected.csv"
-
 
             with open(changes_expected, 'r') as expected, open(changes_output, 'r') as output:
                 expected_strip = expected.read().strip("\n")
